add vllm/distributed package
